There is now only one spritesheet for Module Maze
Module Maze will now use 32x32 sized icons on its screen, scaled similarly to the implementation in iconic. As such, the larger sheets are no longer necessary for the module.
Also, the names used in Modules.meta must exactly match the icon names for LFA compatiblity, as such, apply the same Regex.Replace to the clipboard as reading of the icon names. There is another issue with my implementation of the LFA where apostrophes do not format correctly in the HTML. As such, I've replaced the apostrophes in the file with the ’ character. This is not included in this commit as it's likely very easy to fix in the LFA.